# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API - Carga Pesada",
        "description": "API RESTFul Carga Pesada",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/notes": {
            "get": {
                "tags": [
                    "Note"
                ],
                "summary": "Listar todas as notas",
                "operationId": "13d45fc8b7335b9bb262c654c31c9814",
                "parameters": [
                    {
                        "name": "cidade",
                        "in": "query",
                        "description": "Filtrar por cidade",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de notas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Note"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Note"
                ],
                "summary": "Criar uma nova nota",
                "operationId": "a1ba52eb868ae49c16b4b219c2fe6c15",
                "requestBody": {
                    "description": "Dados da nova nota",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreUpdateNote"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Nota criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Nota criada com sucesso."
                                        },
                                        "note": {
                                            "$ref": "#/components/schemas/Note"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/notes/{id}": {
            "get": {
                "tags": [
                    "Note"
                ],
                "summary": "Obter detalhes de uma nota específica",
                "operationId": "52f58d9f550f35e0a391673bba6a598d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da nota",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes da nota",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Note"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nota não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Nota não encontrada."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Note"
                ],
                "summary": "Atualizar uma nota existente",
                "operationId": "872eaa2ec57b2e5e55a17284a5bf6e0b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da nota a ser atualizada",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Dados da nota a ser atualizada",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreUpdateNote"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Nota atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Note"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nota não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Note"
                ],
                "summary": "Deletar uma nota existente",
                "operationId": "dae4c0fc408e565e17f0dea75900d0e6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da nota a ser deletada",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Nota deletada com sucesso"
                    },
                    "404": {
                        "description": "Nota não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Note": {
                "required": [
                    "id",
                    "descricao",
                    "cidade",
                    "estado",
                    "valor",
                    "viagem_id"
                ],
                "properties": {
                    "id": {
                        "description": "ID único do caminhão (UUID)",
                        "type": "string",
                        "format": "uuid"
                    },
                    "descricao": {
                        "description": "Descrição",
                        "type": "string"
                    },
                    "cidade": {
                        "description": "Cidade",
                        "type": "string"
                    },
                    "estado": {
                        "description": "Estado",
                        "type": "string"
                    },
                    "valor": {
                        "description": "Valor",
                        "type": "number",
                        "format": "float"
                    },
                    "viagem_id": {
                        "description": "ID único da viagem (UUID)",
                        "type": "string",
                        "format": "uuid"
                    }
                },
                "type": "object"
            },
            "StoreUpdateNote": {
                "title": "StoreUpdateNote",
                "required": [
                    "descricao",
                    "cidade",
                    "estado",
                    "valor",
                    "viagem_id"
                ],
                "properties": {
                    "descricao": {
                        "description": "Descrição",
                        "type": "string"
                    },
                    "cidade": {
                        "description": "Cidade",
                        "type": "string"
                    },
                    "estado": {
                        "description": "Estado",
                        "type": "string"
                    },
                    "valor": {
                        "description": "Valor",
                        "type": "number",
                        "format": "float"
                    },
                    "viagem_id": {
                        "description": "ID único da viagem (UUID)",
                        "type": "string",
                        "format": "uuid"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}